Extend creation of synthetic buildings
Multiple synthetic building will be build in cells for which no osm_ids but
but zensus data for households are available. The number of synth buildings
match the building count from zensus data. If no building information is
available, the mean rate of profile/buildings will be used to derive the
number of buildings from the number of profile ids.