Use default values in lieu of a configuration file
Instead of exiting with an error if no configuration file is found, a
warning is raised and default values are assumed.
While a useful feature in and of itself, this is also a short term fix
to make tests importing `egon.data.airflow.dags.pipeline` pass, because
that pulls in code trying to access configuration values. The real fix
for this though is to run the corresponding tests in a correctly set up
environment having a configuration file. Fortunately, this is exactly
how the warning is squelched, so there'll be reminder whenever tests are
run.