Fix usage of `...` and `>>>` in example code
In a real Python console there wouldn't be three dots, because there is
no expression spanning multiple lines. The imports are separate
expressions, so each one is prefixed by a `>>>`. The example is also
treated as a doctest and fails if the output isn't specified correctly,
so this change is necessary to make all tests pass.