Skip example needing a database connection
Just like the previous commit, `>>>` and `...` were not used correctly
in the example, but even after fixing this issue, the test currently
still can't pass, because it would need a connection to a populated
database. Eventually, one might want to set up a minimal test database
as a test context but until this is done, the test has to be skipped.