Temporarily remove the `check` tox environment
This checks the code for formatting issues and other and since the
automated tests have been disabled for a while now, code formatting has
been going downhill steadily and quite a few issues have accumulated.
These will take a while to fix, until which time this environment will
be disabled.