Remove `macos` and `windows` from "tox.ini"
Having these in there means that a simple `tox` invocation will test
against them, even for somebody who wants to test locally on his
machine, which obviously only has one platform. Since currently that's
almost always Linux, that's the platform we keep as a minimal baseline
environment on which to get the tests running again.
In the long run, I'd also like to make the "tox.ini" as platform
agnostic as possible and push all platform specific code into the
workflow definition file.