Add contributor dict with example usage
Similar to the sources dict in metadata.py I added a contributor
dict, so full name and github handle are easily accesible. Import
the contributors to add them to the metadata. It is possible to
update the dict according to the work done in the respective
section. Might need a .copy() to not manipulate the original dict,
I am not sure however.