Restrict `pandas` to versions < 1.4
Pandas versions >= 1.4 need SQLAlchemy versions >= 1.4 to access
databases. Since we constrain SQLAlchemy to versions < 1.4, not
constraining pandas makes the tasks

  - "vg250.to-postgres",
  - "etrago_setup.insert-carriers",
  - "re_potential_areas.insert-data" and
  - "industrial_gas_demand.download-industrial-gas-demand"

fail with an error. The last one fails in pandas, but it's a different
error than the others, so it might not be related to the SQLAlchemy
version.
Regardless, since database access is not an essential feature of pandas,
SQLAlchemy is not one of its dependencies, which means that pandas also
doesn't constrain the SQLAlchemy version, so we have to do so.