Explain Airflow's upper version constraint
Compared with previous versions, Airflow 2.0 contains enough breaking
changes, which are incompatible with how we are using Airflow, to make
migrating to Airflow 2.0 a non-trivial undertaking. Until we manage to
do so, we have to constrain the Airflow's upper version to be strictly
below 2.0.
Both version constraints where introduced in commit b668cdcc9, but only
the lower one was explained in the corresponding commit message.