Add dependencies needing explicit upper limits
These packages would normally not have to be specified because they
would be pulled in by other dependencies. But since the most up to date
versions of these transitive dependencies contain breaking changes, they
are now listed explicitly so that we can put constraints on the upper
versions to block these breaking changes.
The specific problems are as follows: with the advent of version 3.0
WTForms removes the `wtforms.compat` module. Since this module is used
by Apache Airflow in versions below 2.0, we have to make sure to use a
WTForms version below 3.0. This has further implications, because
starting with version 2.1, MarkupSafe removes the
`markupsafe.soft_unicode` module which is used by WTForms<3.0 so
MarkupSafe has to be constrained as well.