Remove superfluous `.insert_into` implementation
`Operator`s instantiated within a `with airflow.DAG()` block are
automatically added to the instantiated `DAG` so the `.insert_into`
implementation is superfluous, as long as all `Dataset`s and `Operator`s
are instantiated with such a block. Which they currently are. If the
need to change this comes up, I can add the feature again but right now
it's safer if it's not present.