Remove the reference to `database.Setup`'s subtask
The `database.Setup` dataset only has one subtask which is the
`database.setup` function. This subtask was exposed directly to serve as
an example, but it seems that everyone directly referenced that instead
of the dataset. So I changed `setup` to reference the dataset instead of
the task which is cleaner and should work the same.