Inline direct dependencies to `tasks`
Most of these where only used once anyway, so putting them in separate
variables doesn't give any benefit besides (maybe) readability. But to
achieve that, the variable definitions would have to be close to the use
site, which they usually were not. So I decided to remove the variables
and inline their definitions at their use sites.
This has the following benefits: first, when looking at a dependency
list, it is immediately apparent which dependency is a `Dataset` and
which dependency is a task. Second, doing this allowed me to spot that
`elec_annual_household_demands_cells` was actually unused, so I could
remove it.