Replace subtask with it's parent dataset
The task at hand is "zensus_mv_grid_districts.mapping" task which is
stored inside the `map_zensus_grid_districts` variable and belongs to
the `ZensusMvGridDistricts` dataset. In the new, cleaned up pipeline,
the variable no longer exists which leads to an error when loading the
pipeline. Using the parent dataset as a dependency fixes that error.
Note that the `MotorizedIndividualTravel` dataset is the only one which
needs to be taken into account here, because the other ones are either
already fixed during the pipeline cleanup, or they are (going to be)
fixed on their respective branches.