Revert "Merge fix #403 into CI branch"
This reverts commit 060c532cbf9d5836337ab643dbe80ed97145a687, reversing
changes made to 30c5b9f71e773367e41cff14663d277c35ef2599.

Since this is a revert of a merge commit, you have to revert this revert
before merging the branch again or commits which where already on the
branch prior to the reverted merge will not be merged again.
See [How to revert a faulty merge][0] for more information.

TL;DR: You have to do a `git revert <this commit's id>` before merging
       your branch again.

[0]: https://github.com/git/git/blob/master/Documentation/howto/revert-a-faulty-merge.txt