Use (parts of) a UUID4 as the `next_etrago_id`
I did a few experiments and using six bytes of the UUID would've been OK
but risky, but using seven bytes didn't lead to collisions for up to
around 1.6e+10 UUIDs in around 3e+5 contiguous ranges with random
lengths from 1e+4 to 1e+5. Using eight bytes would mean the risk of
generating integers which are to big for PostgreSQL's BigInt datatype,
so seven bytes is the best compromise between simplicity and collision
mitigation.