Avoid a bare except
This is due to a "flake8" complaint, but it also makes the code more
explicit. The intention is to provide an alternative, should the
`dataset_2040[node]` lookup fail. Explicitly catching only the
`KeyError` caused by this avoids potentially hiding other errors.
There's still a bug in this code, as explained in issue #926, but that's
something which has to be fixed separately.