Fix potentially dangerous duplicate calls
Calling `next_etrago_id` two times in "close" succession did not
guarantee getting the same results even before switching to random IDs.
This was due to concurrency. Now that we're using random IDs, these two
calls are basically guaranteed to yield wildly different results (which
is the point of the new implementation), so this pattern is very
dangerous and will nearly always be wrong.
Fortunately it is easily fixed by saving the result of the first call
and using it in place of both calls.