Add a failing test for issue #985
In order to reproduce the issue, the test creates two `Dataset`s ending
in parallel tasks and sets their `__module__` attribute to a module
below the `egon.data.datasets` package to simulate having the `Dataset`s
created there. Then the test checks that the ids of the final `Dataset`
tasks are distinct.