Be more detailed in `assign_gas_bus_id`'s docstring
The first line of the docstring was too long, because it was longer than
the absolute limit of 79 characters and because it was longer than the
limit of 72 characters for free flowing text. So I shortened it. And
because shortening was done easiest by moving the remark in parenthesis to
the long description, that's what I did.