Use a single `sqlalchemy.engine.Engine` per process
According to the [SQLAlchemy documentation][0], the `Engine` should be
"held globally", but also "initialized per process". Since parallelizing
the workflow was the whole point of "egon-data" holding the `Engine`
globally wasn't really an option, but I went overboard with an API that
creates a new `Engine` for every `Session`. Fortunately creating the
`Engine` through a factory function allows us to cache the `Engine` on a
per process basis. This should hit the sweet spot demanded by the
[SQLAlchemy documentation][0].

[0]: https://docs.sqlalchemy.org/en/13/core/connections.html#basic-usage