Add a helper converting result rows to dictionaries
The "problem" that this solves is that SQLAlchemy has a weird quirk in
that a `Query` returns data which is structured differently depending on
what is queried. If a single mapped class is queried, the query returns
a list of instances of the mapped class where each instance corresponds
to a row of the query result. This case is a bit harder to convert to
dictionaries, because one has to make use of the `__table__` attribute.
All other cases, i.e. querying multiple mapped classes, explicitly
listing the columns to query or a combination of both, results in a list
of keyed tuples, which are much easier to convert to dictionaries.
The helper implemented here combines both cases.