Implement yielding `session` and a `connection`
This context manager can be used everywhere, where a `session` is
needed to interact with the ORM at the same time as a `connection` to
get more direct access to the database. The `session` and the
`connection` share the same transaction and everything will be properly
committed and closed when exiting the context manager.