Don't use session outside of its context manager
In previous commits, this use of `read_postgis` was replaced with a
combination of `GeoDataFrame` and `DataFrame.from_records`. I couldn't
use the same technique here, because there's no `geom_column` argument
to `read_postgis` which means that I don't know which column to convert
using `to_shape`. While this can probably be figured out, I don't have
the time for now so it's a TODO for later.
So in order to not use the session after it is closed (which is not
strictly wrong, because we only use the `bind` attribute, but it still
leaves the door open to unknown behaviour), I'm replacing the session
with a call to `db.engine()`. Due to the per-process caching of engines,
this doesn't incur additional connections, while it also should be
identical in behaviour to using `session.bind`.