Remove `db.engine()` as default parameter values
Since default parameter values are evaluated at function definition time
and stay with the function for its entire lifetime, they are essentially
the same as module level variables (at least for top level functions,
that is). So `db.engine()` as a default parameter value has the same
problems as `db.engine()` at module level and should be removed
accordingly.
Fortunately removing it is as simple as setting the default parameter
value to `None` and then checking for `None` at the start of the
function body, which is what this commit does.