Remove unclosed `session`s
These `session` are used at or near the top of functions and are never
closed, potentially leaking connections. Using the `session_scoped`
decorator on the functions allows us to get a `session` for the whole
function which is automatically committed and closed at the end of the
function.
Note that one `sessionmaker` import gets removed because it's just
unused.