Manually close `session`s
These sessions are not opened via a context manager and thus have to be
closed manually in order to not potentially leak connections. It might
not be necessary but it's best to be on the safe side.
Also, these are `session` usages which I couldn't somehow refactor to
working with a context manager, so this is the minimal effort to stay on
the safe side w.r.t. connection leakage.