Merge too finely divided transactions
These `with` blocks created two transactions inside functions which
where wrapped in retrying error handlers. This could potentially lead to
always failing retries because committed transactions can not be rolled
back, so errors in the second transactions trigger retries on
unchangeable state. In order to prevent this, it's best to have the
whole function be one transaction.