Replace more uses of `read_sql` and `read_postgis`
This is similar to commit 4950314, except that `session.bind` wasn't
used outside of the `session`'s context manager. So the only "problem"
here is that we don't know whether pandas correctly closes the
connection it uses to query the database. Simply using the query result
removes that uncertainty.
Since everything happens inside the context manager, this means we can
just use the `.all()` method of the established query inside the
`pd.DataFrame.from_records` call.