Use "rich" to format logged exceptions
The package has a particularly useful way of formatting how exceptions
are printed, because it prints the context of the code in which the
exception occurred, as well as the variables currently in scope along
with their values. All that while still keeping things pretty readable,
at least IMHO. Unfortunately it's not enough to simply use rich's
default way of switching on pretty printing of exceptions, because rich
also colors the stack traces and the markup used to get those colors
interferes badly with how Airflow displays logged messages in its web
application. For this reason, it's necessary to change the markup when
formatting stack traces.