Log raised exceptions
These are the two places that exceptions bubble up to: the main process
and the callables which are run in parallel processes by Airflow. This
might miss custom `Operator`s, but there's no (easy) way to use
`logger.catch` automatically on those.
The exceptions which are covered should automatically be formatted using
"rich"'s exception formatting.