Cache `Engine` objects on a per process basis
More generally, implement a function which caches `Engine` objects on a
per argument basis and change `engine` to return the same `Engine` for
every call from the same process, as determined by process id.
Since engines are responsible for managing the connection pool, are
supposed to be created once, globally per application, but are also not
supposed to be shared across process boundaries, this is the best
compromise which hopefully fixes or at least mitigates issue #799.