Bump pre-commit hook versions
This is the result of running `pre-commit autoupdate`. The new version
of the `flake8` pre-commit hook requires a Python version >= 3.8.1, so
it was necessary to add the `language_version: python3.8` line. Since
`egon-data` also supports Python 3.7, this makes sure that a suitable
interpreter version is chosen to run the pre-commit hook.