Change "CHANGELOG.rst" entry
The URL used to link to the corresponding issue was incorrect. It linked
to the PR instead of the issue. This works because GH doesn't really
distinguish between PRs and issues and automatically rewrites URLs
according to what the number points to. But it still wasn't clear
whether the link target was supposed to be the PR or the issue. Now both
are linked, external links are used for readability and the wording is a
tad more verbose and friendly (IMHO).