Remove `.*` from ordered version comparison
Apparently this was never legal as a trailing `.*` is only allowed when
doing ["Version matching"][0] but not when doing an ["Inclusive ordered
comparison"][1]. For some reason the tests only start to complain now.
Maybe some package parsing this version specifier got an update and is
now stricter than before.
Anyway, good that this got caught and is now corrected.

[0]: https://peps.python.org/pep-0440/#version-matching
[1]: https://peps.python.org/pep-0440/#inclusive-ordered-comparison