Add missing space
There was no space after the last "&". I don't know whether this
mattered here, but better safe than sorry.
Also, when joining strings over multiple lines, it's good practice to
put spaces at the beginning of strings because then it's easier to spot
missing spaces. That's why I took the liberty of moving the space after
the first "&" to the next line, too.