Constrain the Sphinx version for building the docs
For Pythons < 3.9, the Sphinx version needs to be constrained. Otherwise
it errors out with:

   TypeError: entry_points() got an unexpected keyword argument 'group'

due to Sphinx issue #10479.
Higher Python versions use Sphinx versions which are so up to date,
they get other errors with our documentation. So all in all, we need the
constraint to not run in any weird errors when building the
documentation for now.

[0]: https://github.com/sphinx-doc/sphinx/issues/10479