Use descriptive variable names
Replace single and dual letter variable names with descriptive ones.
Remove comments as the variable names now serve a purpose comparable to
that of the comments.
Feel free to reintroduce the comments if you feel that they serve an
additional informatory purpose. But please expand them a bit, at least
noting that they are examples of what's happening in specific exemplary
loop iterations. That is, if I understood them correctly. If I'm wrong,
please expand them to prevent that misunderstanding.