Try to make code easier to understand
The usage of `i` to store `DataFrame` column names was very confusing
to me. It correctly suggested (to me) that `i` is an integer index so I
got that `x["foo"][i]` is accessing column `"foo"` of the `DataFrame`
stored in `x` and then pulling out the value at index `i` of the
resulting `Series`. But since I now associated `i` with being an integer
index, the left hand side of `x["foo"][i] = y[i].values.tolist()`, i.e.
`y[i].values.tolist()`, didn't make sense to me, because `y[i]` looks
like it's pulling out the single value of a `Series` or list at an
integer index. But it doesn't. It actually pulls the column `i` out of
the `DataFrame` `y`. That can then be converted to a list and is
assigned to the `DataFrame` cell `x["foo"][i]`.
Figuring this out took me hours, so I decided to change the variable
name to `column` to convey that it's meant to store a column, even if it
is sometimes used to index a row. What is indexed is then made explicit
by using `.at` and `.loc` as appropriate.
Using `.at[row, column]` also serves the purpose of squelching some
instances of

  ```
  SettingWithCopyWarning:
  A value is trying to be set on a copy of a slice from a DataFrame.
  Try using .loc[row_indexer,col_indexer] = value instead

  See the caveats in the documentation: https://pandas.pydata.org/pandas-doc#
  ```

Using `.at[r, c]` is necessary in order to tell `pandas` that we want to
assign a sequence to a single `DataFrame` location. Using `.loc[r, c]`
`pandas` would try to somehow align the sequence with the cell's
neighbourhood which usually results in an error.