Customize the `sphinx-apidoc` templates
Reduce clutter by removing some headings like "Module contents",
"Submodules" and "Subpackages". Also, don't differentiate so strictly
between modules and packages. Treat packages just as modules which are
potentially empty apart from references to sub-modules or sub-packages.
Or treat modules as packages with initialization code but no sub-modules
or sub-packages. Either way, treat them uniformly as much as possible.

Second, use only the last component of the fully qualified module or
package name in headings and TOC entries. The fully qualified name
should always be reconstructable by looking at the path under which the
module's/package's documentation page is accessed.

Last but not least, put an empty line at the bottom of "module.rst_t".
Otherwise the generated rST stubs would not have a newline at the end
the file which would make the pre-commit hooks complain. The empty line
at the template is the easiest way to fix this. The correct way would be
to initialize the `jinja2.Environment` that renders the template with
`keep_trailing_newline=True`, but I don't know how to do that within
`sphinx-apidoc`, so using this workaround instead.