Modify generated API reference
Use "docs/reference/egon.data.rst" instead of "docs/reference/index.rst"
as the `toctree` entry for the API reference for better and direct
access to the sub-modules and packages of `egon.data`. Change the
sidebar entry for the API reference to mention the `egon.data` package
and put the reference for `egon.data`'s contents before the listing of
sub-modules and packages for this one package only.
Remove "docs/reference/index.rst" because it's no longer needed now that
"docs/reference/egon.data.rst" takes it's place.
Remove "docs/reference/egon.rst" because it's empty and not interesting.
It's just an unwanted artifact of running `sphinx-apidoc`.