Fix docstring
First, wrap the first line to be within 72 characters. The first line
should actually be a bit shorter, as it's supposed to be a short,
one-line summary akin to the first line of a git commit message.
Longer details go into the extended description after the first line. As
a courtesy to the reader, the parameter and return value descriptions
can also be complete short sentences.
For the description of the return value, the value doesn't need to be
named. Only the type has to be given.
Fix a typo in the type, which used a comma instead of a period. This
also prompted the change from `pandas.DataFrame` to `pd.DataFrame`
because `pandas` is imported as `pd` which means that `pd.DataFrame` can
actually be linked when the "intersphinx" extensions gets configured
later.