Fix `datetime` usage
There where a lot of `datetime.now()` calls in the module. But `now`
isn't a member of the `datetime` module, but a member of the `datetime`
class inside the `datetime` module.
This is why I don't like modules with members named exactly like the
module. It's an accident waiting to happen.