Make long dictionary string values more readable
Use a parenthesized expression to break long string values in
dictionaries into multiple lines, i.e. use

```python
{
    "first-key": (
        "-----------------------------------------------------"
        "\nA very long string value that has"
        " to be split over multiple lines."
    ),
    "second-key": "Did you spot the second dictionary entry?",
}
```

in favour of

```python
{
    "first-key": "-----------------------------------------------------"
    "\nA very long string value that has "
    "to be split over multiple lines.",
    "second-key": "Did you spot the second dictionary entry?",
}
```

The first version separates keys and values more clearly, making it
easier to spot all keys and to read the string value as a whole. Using
the second version, it's harder to see on which line a new entry with a
new key begins, especially if one doesn't start continuation lines with
a space or other separating character.
See the message of the previous commit for details on why to put spaces
and other important separators at the beginning of continuation lines.

Note also that, while most of the lines were within the 79 character
limit, so not too long, they where reformatted because of the reasons
above. And since they had to be touched anyways, they got wrapped at 72
characters, as that's what PEP8 suggests for free flowing text like
comments and long strings.