Wrap long strings to 72 characters
This is the length that PEP8 states for free flowing text like comments
or long strings. Since most strings don't start at the beginning of the
line, I employed some leeway to go a few characters over in favour of
readability and sometimes I wrapped lines early for the same reason,
mostly at sentence boundaries.
See the previous two commits for details on this particular way of
wrapping strings.