Invert `file.endswith(".json")` test
Use this to save an indentation level. The rest of the code after the
`continue` is exactly the same as before, just indented one level less.
This isn't really important now, but since the code will change to get
rid of the test altogether, it's a precursor to not mixing getting rid
of an indentation level and other code changes into one commit later on.