Make logging message more precise
It's not the table that gets "uploaded", but the metadata comment. And
things don't (always) get "uploaded" to a database, but they get stored
in the database, because the database might be on the same machine.
Last but not least, there's no reason to use use an exclamation mark in
log messages. It just makes long streams of log messages harder to
read, if they aren't terminated consistently. Log levels usually are
enough to signify the importance of message.