Move JSON metadata files to the "metadata" package
Get rid of the hardcoded "json_metadata" directory name and instead use
`__name__` to get the name of the current module, which now is also a
package, then use `importlib_resources.files` to get a `Path` to the
files inside the package, then use `.glob("*.json")` to get all the
files ending with ".json" below that path.

This simplifies the code because we no longer have to:

  - import `Path`, `os` or `egon.data`'s `__path__`,
  - generate the path to the files by hardcoding their directory,
  - filter for files ending in ".json" and
  - manually prefix the filenames with the path under which they where
    found.

There is one important change though: since `.glob("*.json")` returns a
generator of `Path` objects, the filename component has to be pulled out
of the `Path` via the `name` attribute before `split`ting on `"."`.

Last but not least, add a missing newline at the end of
"src/egon/data/datasets/zensus_vg250.py", courtesy of the
`end-of-file-fixer` pre-commit hook.