Remove spurious erroneous line
The line only contains the word `node` which breaks the Python file and
all other files trying to import it, because Python thinks that's an
identifier and errors out, since the name isn't defined anywhere.
It's probably just a copy paste error or a typo that snuck into the
commit.