Wrap docstring lines at 72ish characters
[PEP 8][0] actually specifies a maximum line length of 72 characters for
"flowing long blocks of text with fewer structural restrictions
(docstrings or comments)". I extended that to 76 characters for indented
docstrings, because the reason for having 72 characters is so that the
eyes don't have to travel horizontally too much. Indentation basically
moves the start of the text block to the right, so for text blocks like
this, the line length limit can move with the indentation.
Some of these lines where even beyond 79 characters, so I had to touch
them anyways to make `flake8` happy and while at it, I also changed them
to be (mostly) PEP 8 compliant.

[0]: https://peps.python.org/pep-0008/#maximum-line-length