Replace `partial`s of `Dataset` with subclasses
Make these subclasses `@dataclass`es, just like the original `Dataset`,
so that default values for the fields are picked up by the constructor.
That way we don't have to specify a constructor that calls the
superclass constructor, but can rely on the default implementation for
`__init__`.
The `dependencies=[]` line present in nearly all uses of `partial` has
been removed because it was unnecessary. The `dependencies` parameter is
usually specified when the datasets are instantiated and even if not, an
empty list of dependencies is the default anyways. Actually, an empty
tuple is the default, but that doesn't make a difference.
The `tasks=(define_mv_grid_districts)` line has been changed to drop the
parenthesis because the parenthesis where superfluous in this case.
The `tasks=(download_mastr_data,)` line has not been changed, because
even though `Dataset`s should handle a callable and a tuple
containing exactly one callable identically for the tasks attribute, I
didn't want to risk introducing any bugs.