Manually specify `Dataset` subclass constructors
Instead of relying on `@dataclass` to provide one. This makes the
constructor's signature less confusing. I also allows not specifying
types for the class variables, because `@dataclass` no longer needs to
pick the up automatically, as we are manually specifying them in the
superclass constructor call.
This means we have less stuff to import, most notably we no longer need
to import `Tasks`. This also allows us to work around the broken display
of `mv_grid_districts_setup`'s `tasks`. That was only a single function
and for some reason Sphinx did not display that one correctly. Probably
because it tried to use `str` instead of `repr` to render the function,
because boxing the function in a one-tuple would fix the display.
Anyway. Not putting the task on a class attribute but specifying it
directly in the constructor means that no class attribute is displayed
in the documentation, thus sidestepping the issue.