Fill empty columns with np.nan to allow multiplication
If values would be none, an error would be raised. In this case
multiplication is possible but results in nan as well which is
not a problem as this scenario is not selected in the run but
due to the code needs to be calculated anyways. This is more an
adhoc fix. A clean fix would replace the specifc factor_* columns
and make them dynamic or organise in a different way.