Replace Centroid of industrial area with PointOnSurface of the industrial area
There are a few edge cases where the Centroid of the area is not inside the area and
even outside of the mv grid districts because it is e.g. in the sea (at harbors).
In those cases, no bus_id can be assigned. To prevent these problems, a point on the
surface is used instead.