Create distrct heating profiles with census cells selected in annual demand
A few cells are assigned to two district heating grids. This is a bug.
To allow the correct creation of the profiles, the celss are siggned to one of the
grids. In the longer term, the creation of district heating areas needs to be fixed.
When this is done, this commit can be reverted.